Refactored out each section in the app to its own controller.
The Header is now its own controller(Contains no code for now)

The CustomerNameList has been extracted:

I created an event for MoveLocation to handle the swapping motion on the main page because the overall layout is the responsibility of the main page, not the individual components.

Additionally I created an event that updates the model displayed in the customer details panel because I didn't want any coupling between the separate components. Instead the main page handles  any changes.